new_bot/requirements.txt
aiogram>=3.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pydantic>=2.0.0
greenlet>=3.0.0
